Small changes to config and custom auth provider.

+++ src/main/java/com/delacasa/auth/config/AppLoginConfig.java
@@ -16,6 +16,6 @@ public class AppLoginConfig {
 	private final String successUrl;
 	private final String twoFAuthRequiredUrl;
 	private final String mailRegex;
-	private final byte maxTries;
+	private final int maxTries;
 
 }
